feat: MainActivity 설정
