+++ presentation/src/main/AndroidManifest.xml
@@ -12,6 +12,16 @@
         android:label="@string/app_name"
         android:theme="@style/Theme.Duckie.Splash">
 
+        <activity
+            android:name=".MainActivity"
+            android:exported="true">
+
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
+        </activity>
+
     </application>
 
 </manifest>
