--- presentation/src/main/kotlin/team/duckie/app/android/presentation/CrashActivity.kt
@@ -1,33 +1,0 @@
-/*
- * Designed and developed by Duckie Team, 2022
- *
- * Licensed under the MIT.
- * Please see full license: https://github.com/duckie-team/duckie-android/blob/develop/LICENSE
- */
-
-package team.duckie.app.android.presentation
-
-import android.app.Activity
-import android.os.Bundle
-import android.view.ViewGroup
-import android.widget.Button
-
-class CrashActivity : Activity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        val crashButton = Button(this)
-        crashButton.text = getString(R.string.app_name)
-        crashButton.setOnClickListener {
-            throw RuntimeException("Test Crash") // Force a crash
-        }
-
-        addContentView(
-            crashButton,
-            ViewGroup.LayoutParams(
-                ViewGroup.LayoutParams.MATCH_PARENT,
-                ViewGroup.LayoutParams.WRAP_CONTENT,
-            ),
-        )
-    }
-}
