+++ presentation/src/main/kotlin/team/duckie/app/android/presentation/MainActivity.kt
@@ -1,0 +1,62 @@
+/*
+ * Designed and developed by Duckie Team, 2022
+ *
+ * Licensed under the MIT.
+ * Please see full license: https://github.com/duckie-team/duckie-android/blob/develop/LICENSE
+ */
+
+package team.duckie.app.android.presentation
+
+import android.animation.ObjectAnimator
+import android.os.Build
+import android.os.Bundle
+import android.view.View
+import android.view.ViewGroup
+import android.view.animation.AnticipateInterpolator
+import android.widget.Button
+import androidx.activity.ComponentActivity
+import androidx.core.animation.doOnEnd
+import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
+import androidx.core.view.WindowCompat
+
+private const val DefaultSplashScreenExitAnimationDurationMillis = 200L
+
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        installSplashScreen()
+        WindowCompat.setDecorFitsSystemWindows(window, false)
+        super.onCreate(savedInstanceState)
+
+        val crashButton = Button(this)
+        crashButton.text = getString(R.string.app_name)
+        crashButton.setOnClickListener {
+            throw RuntimeException("Test Crash") // Force a crash
+        }
+
+        addContentView(
+            crashButton,
+            ViewGroup.LayoutParams(
+                ViewGroup.LayoutParams.MATCH_PARENT,
+                ViewGroup.LayoutParams.MATCH_PARENT,
+            ),
+        )
+
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S) {
+            splashScreen.setOnExitAnimationListener { splashScreenView ->
+                ObjectAnimator.ofFloat(
+                    splashScreenView,
+                    View.ALPHA,
+                    1f,
+                    0f,
+                ).run {
+                    interpolator = AnticipateInterpolator()
+                    duration = DefaultSplashScreenExitAnimationDurationMillis
+                    doOnEnd {
+                        splashScreenView.remove()
+                    }
+                    start()
+                }
+            }
+        }
+    }
+}
